Crate bio_types

Modules

alignment

Types for representing pairwise sequence alignments

annot

Data types for positions and regions on named sequences (e.g. chromosomes), useful for annotating features in a genome. For example, these data types let you represent that TMA22 is on chromosome X, positions 461,829-462,426, on the forward strand. They also allow coordinate math on these annotations, e.g., that position chrX:461,839 is +10 within TMA22 and vice versa.

genome

sequence

strand

Data types for strand information on annotations.

variant