Struct bio_types::annot::contig::Contig

pub struct Contig<R, S> { /* fields omitted */ }

Contiguous sequence region on a particular, named sequence (e.g. a chromosome)

Parameterized over the type of the reference sequence identifier and over the strandedness of the position.

The display format for a Contig is chr:start-end(+/-/.). The boundaries are given as a half-open 0-based interval, like the Rust Range and BED format.

use bio_types::annot::contig::Contig;
use bio_types::strand::ReqStrand;
let tma19 = Contig::new("chrXI".to_owned(), 334412, (334916 - 334412), ReqStrand::Reverse);
let tma19_str = tma19.to_string();
assert_eq!(tma19_str, "chrXI:334412-334916(-)");
let tma19_str_loc = tma19_str.parse()?;
assert_eq!(tma19, tma19_str_loc);

Implementations

impl<R, S> Contig<R, S>

pub fn new(refid: R, start: isize, length: usize, strand: S) -> Self

Construct a new sequence contig location

use std::rc::Rc;
use bio_types::annot::contig::Contig;
use bio_types::strand::ReqStrand;
let chr = Rc::new("chrX".to_owned());
let tma22 = Contig::new(chr, 461829, 462426 - 461829, ReqStrand::Forward);

pub fn with_first_length(
pos: &Pos<R, S>,
length: usize
) -> Result<Self, AnnotError> where
R: Clone,
S: Into<Option<ReqStrand>> + Copy,

Construct a new sequence contig location from a starting position and length.

In general, the starting position must have a "strandedness", and reverse-strand starting positions will extend towards lower coordinates from the starting position.

use bio_types::annot::contig::Contig;
use bio_types::annot::pos::Pos;
use bio_types::strand::ReqStrand;

let tma22_first = Pos::new("chrX".to_string(), 461829, ReqStrand::Forward);
let tma22 = Contig::with_first_length(&tma22_first, 462426 - 461829)?;
assert_eq!(tma22.to_string(), "chrX:461829-462426(+)");

let tma19_first = Pos::new("chrXI".to_string(), 335015, ReqStrand::Reverse);
let tma19 = Contig::with_first_length(&tma19_first, 335016 - 334412)?;
assert_eq!(tma19.to_string(), "chrXI:334412-335016(-)");

pub fn into_stranded(self, strand: ReqStrand) -> Contig<R, ReqStrand>

Convert into a stranded sequence location on the specified strand

impl<R> Contig<R, ReqStrand>

pub fn extend_upstream(&mut self, dist: usize)

Extend the annotation by dist in the upstream direction on the annotated strand.

Arguments

dist specifies the offset for sliding the position. The left, 5'-most end of the contig will expand for forward-strand annotations and the right, 3'-most end will expand for reverse-strand annotations.

use bio_types::annot::contig::Contig;
use bio_types::strand::ReqStrand;
let mut tma22 = Contig::new("chrX".to_owned(), 461829, 462426 - 461829, ReqStrand::Forward);
tma22.extend_upstream(100);
assert_eq!(tma22.to_string(), "chrX:461729-462426(+)");
let mut tma19 = Contig::new("chrXI".to_owned(), 334412, 334916 - 334412, ReqStrand::Reverse);
tma19.extend_upstream(100);
assert_eq!(tma19.to_string(), "chrXI:334412-335016(-)");

pub fn extend_downstream(&mut self, dist: usize)

Extend the annotation by dist in the downstream direction on the annotated strand.

Arguments

dist specifies the offset for sliding the position. The right, 3'-most end of the contig will expand for forward-strand annotations and the left, 5'-most end will expand for reverse-strand annotations.

use bio_types::annot::contig::Contig;
use bio_types::strand::ReqStrand;
let mut tma22 = Contig::new("chrX".to_owned(), 461829, 462426 - 461829, ReqStrand::Forward);
tma22.extend_downstream(100);
assert_eq!(tma22.to_string(), "chrX:461829-462526(+)");
let mut tma19 = Contig::new("chrXI".to_owned(), 334412, 334916 - 334412, ReqStrand::Reverse);
tma19.extend_downstream(100);
assert_eq!(tma19.to_string(), "chrXI:334312-334916(-)");