Struct bio_types::annot::pos::Pos

pub struct Pos<R, S> { /* private fields */ }

Position on a particular, named sequence (e.g. a chromosome).

Parameterized over the type of the reference sequence identifier and over the strandedness of the position.

The display format for a Pos is chr:pos(+/-). A stranded position must have a (+) or a (-), while an unstranded position does not.

use bio_types::annot::pos::Pos;
use bio_types::strand::ReqStrand;
let start = Pos::new("chrIV".to_owned(), 683946, ReqStrand::Reverse);
let start_str = start.to_string();
assert_eq!(start_str, "chrIV:683946(-)");
let start_str_pos = start_str.parse()?;
assert_eq!(start, start_str_pos);

Implementations

impl<R, S> Pos<R, S>

pub fn new(refid: R, pos: isize, strand: S) -> Self

Construct a new sequence position

use std::rc::Rc;
use bio_types::annot::pos::Pos;
use bio_types::strand::ReqStrand;
let chr = Rc::new("chrIV".to_owned());
let start = Pos::new(chr, 683946, ReqStrand::Reverse);

pub fn pos(&self) -> isize

Position on the reference sequence (0-based).

pub fn into_stranded(self, strand: ReqStrand) -> Pos<R, ReqStrand>

Convert into a stranded sequence position on the specified strand